backend/requirements.txt
flask
torch
torchvision
torchaudio
captum
transformers
Pillow
matplotlib
numpy
scikit-learn
timm
